Moved Scheduler state files into a singular file
